backend/requirements.txt
Flask==3.0.2
pymongo==4.6.2
PyJWT==2.8.0
flask-cors==4.0.0